Add files via upload
An initial working version of an LSTM stock prediction model.
There needs to be confirmation on the sizes of train, test, val datasets when stacked. 
Additionally, it has been suggested (and I agree) that we might consider cutting the testing data to a more manageable size for the model to predict (reduce the predicted time frame). This can be done in the utility file, when loading the data, by increasing "trainPercent" I might try increasing it to 80-90% and reduce validation accordingly.